Make Pro + kiosk live-only, unify mode naming, and upgrade to Opus 4.8.
Aligns the marketing (/) and kiosk surfaces with the /showcase golden source:
removes all fixture/demo data so every surface renders live Aurora data, renames
modes to SQL/MCP/Retrieval/Production/Workflow (no phase numbers anywhere user-
facing), swaps the default Bedrock model to global.anthropic.claude-opus-4-8, and
polishes the SQL no-results reply through the concierge LLM path. Also fixes the
seed_data module import + stale loopback proxy handling.

+++ meridian/backend/agents/retrieval_03/search_agent.py
@@ -155,8 +155,11 @@ When searching:
         search_start = datetime.utcnow()
         candidate_limit = max(limit * config.search.rerank_candidate_multiplier, 25)
 
+        # Cast both args explicitly. Without ::integer, Python ints arrive as
+        # bigint and Postgres can't find a matching overload (the function is
+        # declared as semantic_trip_search(vector, integer)).
         sql = """
-            SELECT * FROM semantic_trip_search(%s::vector, %s)
+            SELECT * FROM semantic_trip_search(%s::vector, %s::integer)
         """
 
         embedding_str = '[' + ','.join(map(str, query_embedding)) + ']'
